refactored timeline routes and components
